backend/requirements.txt
flask
pymysql
requests
openai
jwt
requests
pymysql-pooling